Move editor api back to editor
