+++ spyder/plugins/editor/api/editorextension.py
