+++ spyder/plugins/editor/panels/classfunctiondropdown.py
@@ -13,7 +13,7 @@ from qtpy.QtCore import QSize, Qt, Slot
 from qtpy.QtWidgets import QComboBox, QHBoxLayout
 
 # Local imports
-from spyder.api.panel import Panel
+from spyder.plugins.editor.api.panel import Panel
 from spyder.config.base import _
 from spyder.plugins.completion.manager.api import SymbolKind
 from spyder.utils import icon_manager as ima
