+++ spyder/plugins/editor/panels/codefolding.py
@@ -30,7 +30,7 @@ from qtpy.QtGui import (QTextBlock, QColor, QFontMetricsF, QPainter,
 
 # Local imports
 from spyder.plugins.editor.api.decoration import TextDecoration, DRAW_ORDERS
-from spyder.api.panel import Panel
+from spyder.plugins.editor.api.panel import Panel
 from spyder.plugins.editor.utils.editor import (TextHelper, DelayJobRunner,
                                                 drift_color)
 import spyder.utils.icon_manager as ima
