+++ spyder/plugins/editor/widgets/codeeditor.py
@@ -49,10 +49,10 @@ from spyder_kernels.utils.dochelpers import getobj
 
 
 # Local imports
-from spyder.api.panel import Panel
 from spyder.config.base import _, get_debug_level, running_under_pytest
 from spyder.config.manager import CONF
 from spyder.plugins.editor.api.decoration import TextDecoration
+from spyder.plugins.editor.api.panel import Panel
 from spyder.plugins.editor.extensions import (CloseBracketsExtension,
                                               CloseQuotesExtension,
                                               DocstringWriterExtension,
